Made all tests passing.
The twisted.unittest package keeps a list of all exceptions, whether they are caught or not.
You have to explicitly flush them to get rid of problems. This normally happens when you use assertRaises. Since we're using the yields, we have to manually clean them. The added code does that so that all tests now show a beautiful green OK!